pda-backend/requirements.txt
connexion[swagger-ui]
Flask[async]
flask-socketio
simple-websocket
flask-cors
alpha_vantage
pytankerkoenig
pytest
pytest-cov
pytest-mock
pytest-func-cov
pytest-asyncio
apscheduler
kink
geopy
jokeapi
sendgrid
spotipy
requests
icalendar
tmdbv3api
JustWatch
python-dotenv
googlemaps
pillow
cachetools
